+++ src/main/java/jamdoggie/betterbattletowers/BattleTowerMod.java
@@ -7,7 +7,7 @@ import jamdoggie.betterbattletowers.entity.MobAgressiveZombiePig;
 import jamdoggie.betterbattletowers.entity.golem.GolemVariants;
 import jamdoggie.betterbattletowers.entity.golem.MobGolem;
 import jamdoggie.betterbattletowers.block.TileEntityChestTower;
-import jamdoggie.betterbattletowers.worldgen.data.loot.LootTables;
+import jamdoggie.betterbattletowers.worldgen.data.loot.LootLoader;
 import jamdoggie.betterbattletowers.worldgen.structures.WorldFeatureReverseTower;
 import jamdoggie.betterbattletowers.worldgen.structures.WorldFeatureVanquishedTower;
 import jamdoggie.betterbattletowers.worldgen.data.decoration.Tower;
@@ -30,7 +30,7 @@ import turniplabs.halplibe.util.*;
 
 import java.util.Random;
 
-import static jamdoggie.betterbattletowers.worldgen.data.loot.LootTables.LAPIZ;
+import static jamdoggie.betterbattletowers.worldgen.data.loot.LootLoader.LAPIZ;
 import static net.minecraft.core.net.command.util.CommandHelper.registerWorldFeatureClass;
 
 
@@ -81,7 +81,7 @@ public class BattleTowerMod implements ModInitializer, GameStartEntrypoint, Reci
 	@Override
 	public void afterGameStart() {
 		Tower.init();
-		LootTables.init();
+		LootLoader.init();
 		GolemVariants.init();
 	}
