+++ src/main/java/jamdoggie/betterbattletowers/worldgen/data/defaults/LootTableDefault.java
@@ -3,7 +3,7 @@ package jamdoggie.betterbattletowers.worldgen.data.defaults;
 import jamdoggie.betterbattletowers.block.BattleTowerBlocks;
 import jamdoggie.betterbattletowers.worldgen.data.loot.LootEntry;
 import jamdoggie.betterbattletowers.worldgen.data.loot.LootTable;
-import jamdoggie.betterbattletowers.worldgen.data.loot.LootTables;
+import jamdoggie.betterbattletowers.worldgen.data.loot.LootLoader;
 import net.minecraft.core.WeightedRandomLootObject;
 import net.minecraft.core.block.Blocks;
 import net.minecraft.core.item.IItemConvertible;
@@ -88,13 +88,13 @@ public class LootTableDefault {
 		List<LootEntry> l7 = new ArrayList<>();
 		l7.add(looObj(Items.DUST_REDSTONE, 0, 1.0f, 3, 5));
 		l7.add(looObj(Items.OLIVINE, 0, 1.0f, 2, 4));
-		l7.add(looObj(Items.DYE, LootTables.LAPIZ, 1.0f, 1, 2));
+		l7.add(looObj(Items.DYE, LootLoader.LAPIZ, 1.0f, 1, 2));
 		l7.add(looObj(Items.QUARTZ, 0, 1.0f, 1, 2));
 		lootTables.add(table(index++, l7));
 
 		List<LootEntry> l8 = new ArrayList<>();
 		l8.add(looObj(Items.OLIVINE, 0, 1.0f, 2, 4));
-		l8.add(looObj(Items.DYE, LootTables.LAPIZ, 1.0f, 1, 2));
+		l8.add(looObj(Items.DYE, LootLoader.LAPIZ, 1.0f, 1, 2));
 		l8.add(looObj(Items.QUARTZ, 0, 1.0f, 1, 2));
 		l8.add(looObj(Items.CHAINLINK, 0, 1.0f, 2, 3));
 		l8.add(looObj(Items.BUCKET_LAVA, 0, 1.0f, 1));
@@ -102,7 +102,7 @@ public class LootTableDefault {
 		lootTables.add(table(index++, l8));
 
 		List<LootEntry> l9 = new ArrayList<>();
-		l9.add(looObj(Items.DYE, LootTables.LAPIZ, 1.0f, 1, 2));
+		l9.add(looObj(Items.DYE, LootLoader.LAPIZ, 1.0f, 1, 2));
 		l9.add(looObj(Items.CHAINLINK, 0, 1.0f, 2, 3));
 		l9.add(looObj(BattleTowerBlocks.PRISON_BAR, 0, 1.0f, 2, 3));
 		l9.add(looObj(Items.DIAMOND, 0, 1.0f, 1));
