+++ src/main/java/jamdoggie/betterbattletowers/worldgen/data/loot/LootLoader.java
@@ -37,7 +37,7 @@ import java.util.*;
 import static jamdoggie.betterbattletowers.BattleTowerMod.LOGGER;
 import static jamdoggie.betterbattletowers.BattleTowerMod.MOD_ID;
 
-public class LootTables {
+public class LootLoader {
 	public static final int LAPIZ = DyeColor.BLUE.itemMeta;
 	public static final double INC = 1.20f; // 1.29f
 	public static final int MAX_TIER = 9;
@@ -47,7 +47,7 @@ public class LootTables {
 	public static final String ASSETS_JAR_PATH = String.format("/assets/%s/tower/loot.conf", MOD_ID, MOD_ID);
 	private static boolean init = false;
 
-	private LootTables() {
+	private LootLoader() {
 	}
 
 
@@ -66,9 +66,9 @@ public class LootTables {
 		Type type = new TypeToken<List<LootTable>>() {
 		}.getType();
 
-		if (Files.notExists(LootTables.CONFIG_PATH)) {
+		if (Files.notExists(LootLoader.CONFIG_PATH)) {
 			try {
-				if(FabricLoader.getInstance().isDevelopmentEnvironment() && DataLoader.class.getResourceAsStream(LootTables.ASSETS_JAR_PATH) == null){
+				if(FabricLoader.getInstance().isDevelopmentEnvironment() && DataLoader.class.getResourceAsStream(LootLoader.ASSETS_JAR_PATH) == null){
 					List<LootTable> defaultTable = LootTableDefault.getDefaultTable();
 					String json = gson.toJson(defaultTable, type);
 					Config config = ConfigFactory.parseString("{ loot_tables = " + json + " }");
@@ -82,30 +82,30 @@ public class LootTables {
 						.getBytes(StandardCharsets.UTF_8));
 					Files.copy(devPath, CONFIG_PATH, StandardCopyOption.REPLACE_EXISTING);
 				}else{
-					try (InputStream stream = DataLoader.class.getResourceAsStream(LootTables.ASSETS_JAR_PATH)) {
+					try (InputStream stream = DataLoader.class.getResourceAsStream(LootLoader.ASSETS_JAR_PATH)) {
 						if (stream == null) {
 							LOGGER.error("LootTable could not be regenerated, betterbattletowers.jar is corrupted.");
-							throw new RuntimeException("Asset not found in JAR: " + LootTables.ASSETS_JAR_PATH);
+							throw new RuntimeException("Asset not found in JAR: " + LootLoader.ASSETS_JAR_PATH);
 						}
-						Files.copy(stream, LootTables.CONFIG_PATH);
+						Files.copy(stream, LootLoader.CONFIG_PATH);
 					}
 				}
 			} catch (IOException e) {
 				throw new RuntimeException("Failed to create default loot config at: "
-					+ LootTables.CONFIG_PATH.toAbsolutePath(), e);
+					+ LootLoader.CONFIG_PATH.toAbsolutePath(), e);
 			}
 		}
 
 		try {
-			Config config = ConfigFactory.parseFile(LootTables.CONFIG_PATH.toFile());
+			Config config = ConfigFactory.parseFile(LootLoader.CONFIG_PATH.toFile());
 			String jsonString = config.getValue("loot_tables")
 				.render(ConfigRenderOptions.concise());
 			List<LootTable> table = gson.fromJson(jsonString, type);
-			LootTables.createTables(table);
+			LootLoader.createTables(table);
 		} catch (ConfigException e) {
-			throw new RuntimeException("Failed to parse HOCON config: " + LootTables.CONFIG_PATH.toAbsolutePath(), e);
+			throw new RuntimeException("Failed to parse HOCON config: " + LootLoader.CONFIG_PATH.toAbsolutePath(), e);
 		} catch (JsonSyntaxException e) {
-			throw new RuntimeException("Loot configuration file is malformed: " + LootTables.CONFIG_PATH.toAbsolutePath(), e);
+			throw new RuntimeException("Loot configuration file is malformed: " + LootLoader.CONFIG_PATH.toAbsolutePath(), e);
 		}
 	}
 
@@ -143,7 +143,7 @@ public class LootTables {
 		if (inventory == null) return;
 		List<ItemStack> stacks = generateLootList(random, tier, lootAmount);
 		for (ItemStack stack : stacks) {
-			LootTables.placeItemInChest(random, stack, inventory);
+			LootLoader.placeItemInChest(random, stack, inventory);
 		}
 	}
